Make config interventions only when authhttp is the active auth plugin
authhttp needs to do some config interventions in order to ensure a consistent user
experience. For example, the "login" action gets disabled because it does not make
sense with HTTP authentication.

Previously these interventions were ALWAYS active, even when authhttp had just been
activated in the plugin manager but wasn't the active authentication plugin. Now we
check so beforehand, supporting both the use of authhttp on its own as well as
the primary auth plugin when used together with authsplit.